Refactor stuff into dedicated package
