+++ robots/cmd/test-report/BUILD.bazel
@@ -5,13 +5,16 @@ go_library(
     srcs = ["main.go"],
     embedsrcs = [
         "test-report.gohtml",
+        "compute-config.yaml",
         "default-config.yaml",
+        "network-config.yaml",
+        "storage-config.yaml",
     ],
     importpath = "kubevirt.io/project-infra/robots/cmd/test-report",
     visibility = ["//visibility:private"],
     deps = [
         "//robots/pkg/flakefinder:go_default_library",
-        "//robots/pkg/jenkins:go_default_library",
+        "//robots/pkg/test-report:go_default_library",
         "@com_github_bndr_gojenkins//:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@com_github_spf13_cobra//:go_default_library",
@@ -30,5 +33,8 @@ go_test(
     srcs = ["main_test.go"],
     data = glob(["testdata/**"]),
     embed = [":go_default_library"],
-    deps = ["@com_github_sirupsen_logrus//:go_default_library"],
+    deps = [
+        "//robots/pkg/test-report:go_default_library",
+        "@com_github_sirupsen_logrus//:go_default_library",
+    ],
 )
