+++ robots/pkg/test-report/BUILD.bazel
@@ -1,0 +1,19 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["test_report.go"],
+    importpath = "kubevirt.io/project-infra/robots/pkg/test-report",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//robots/pkg/jenkins:go_default_library",
+        "@com_github_bndr_gojenkins//:go_default_library",
+        "@com_github_sirupsen_logrus//:go_default_library",
+    ],
+)
+
+go_test(
+    name = "go_default_test",
+    srcs = ["test_report_test.go"],
+    embed = [":go_default_library"],
+)
